V80 Updates
**Jabberwocky V80 UPDATE**

The motor mount and the mid body of the extruder has some new geometries introduce to make installation easier. This now excludes the need for the M5 spacers that were previously part of the bearing stack for the drive idler gear. This will make installation much easier and ensure alignment of the drive gears. Both parts now have plastic standoffs built into the parts and will no longer accept the M5 spacers in them. A new bearing installation too has also be updated for the bearing installation.

Prior to this, it was identified that in a small sample of cases some tolerance issues may introduce misalignments as well as the spacers installed incorrectly. It is recommended to replace the two parts with the latest V80 revision to avoid any future issues.

This update is also part of the upcoming  LDO Jabberwocky Assembled kit.